[cli] Move peergos.server.simulator.Simulator to  peergos.server.tests.Simulator
